+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `fetch-portable-object` command.
+- Enabled full text search (limited to posts created by the current user).
 
 ### Changed